refactor: split `reparameterisations` into a submodule
